Reduce the amount of headers needed by the dynamic scopes at runtime: * Make GenericValue, clang::QualType opaque in Value * Move LifetimeHandler implementation into source; requires separate header *
Also:
* Value: replace CPP macros by overloads


git-svn-id: http://root.cern.ch/svn/root/trunk@48675 27541ba8-7e3a-0410-8455-c3a389f83636

+++ include/cling/Interpreter/StoredValueRef.h
@@ -26,7 +26,7 @@ namespace cling {
 
       /// \brief Number of bytes that need to be allocated to hold a value
       /// of our type
-      int64_t getAllocSizeInBytes(const clang::ASTContext& ctx) const;
+      long long getAllocSizeInBytes(const clang::ASTContext& ctx) const;
 
       /// \brief Memory allocated for value, owned by this value
       ///
